+++ tests/CLIFramework/ArgumentInfoListTest.php
@@ -1,20 +1,20 @@
 <?php
-use CLIFramework\ArgumentInfo;
-use CLIFramework\ArgumentInfoList;
+use CLIFramework\ArgInfo;
+use CLIFramework\ArgInfoList;
 
-class ArgumentInfoListTest extends PHPUnit_Framework_TestCase
+class ArgInfoListTest extends PHPUnit_Framework_TestCase
 {
     public function test()
     {
-        $arguments = new ArgumentInfoList;
+        $arguments = new ArgInfoList;
         ok($arguments);
 
         $a1 = $arguments->add('x');
-        ok( $a1 instanceof ArgumentInfo ); 
+        ok( $a1 instanceof ArgInfo ); 
         is('x' , $a1->name );
 
         $a2 = $arguments->add('y');
-        ok( $a2 instanceof ArgumentInfo ); 
+        ok( $a2 instanceof ArgInfo ); 
         is('y' , $a2->name );
 
         ok( $arguments[0] );