+++ tests/TestApp/Command/ArginfoCommand.php
@@ -6,10 +6,10 @@ use Exception;
 class ArginfoCommand extends Command
 {
 
-    public function arginfo() {
-        $this->arg('name');
-        $this->arg('email');
-        $this->arg('phone')->optional();
+    public function arguments($args) {
+        $args->add('name');
+        $args->add('email');
+        $args->add('phone')->optional();
     }
 
     public function execute($name, $email, $phone = null) { }